fix jest on windows & fix dev server on windows & fix consts.ts & YA-14 fix стилей mui
