+++ packages/server/package.json
@@ -5,13 +5,14 @@
   "scripts": {
     "build": "tsc --p ./tsconfig.prod.json",
     "preview": "node ./dist/index.js",
-    "dev": "NODE_ENV=development nodemon index.ts",
+    "dev": "cross-env NODE_ENV=development nodemon index.ts",
     "lint": "eslint .",
     "format": "prettier --write .",
     "test": "jest ."
   },
   "dependencies": {
     "cors": "2.8.5",
+    "cross-env": "7.0.3",
     "dotenv": "16.0.2",
     "eslint-config-prettier": "8.5.0",
     "express": "4.18.1",
